requirements.txt
pytorch-lightning == 1.2.8
easydict==1.9
torchtext==0.9.0
torch == 1.8.0
PyYAML
pytest==6.2.2
pandas
torchmetrics
datasets==1.8.0
transformers==4.7.0